requirements.txt
django-crispy-forms
crispy-bootstrap5
django-debug-toolbar
django-autocomplete-light
Pillow